+++ src/Settings/Pages/SCIPageReels.m
@@ -18,7 +18,8 @@
                     [SCISetting menuCellWithTitle:SCILocalized(@"p_reels_tap_t") subtitle:SCILocalized(@"p_reels_tap_s") menu:[SCITweakSettings menus][@"reels_tap_control"]],
                     [SCISetting switchCellWithTitle:SCILocalized(@"p_reels_scrubber_t") subtitle:SCILocalized(@"p_reels_scrubber_s") defaultsKey:@"reels_show_scrubber"],
                     [SCISetting switchCellWithTitle:SCILocalized(@"p_reels_unmute_t") subtitle:SCILocalized(@"p_reels_unmute_s") defaultsKey:@"disable_auto_unmuting_reels" requiresRestart:YES],
-                    [SCISetting switchCellWithTitle:SCILocalized(@"p_reels_refresh_t") subtitle:SCILocalized(@"p_reels_refresh_s") defaultsKey:@"refresh_reel_confirm"]
+                    [SCISetting switchCellWithTitle:SCILocalized(@"p_reels_refresh_t") subtitle:SCILocalized(@"p_reels_refresh_s") defaultsKey:@"refresh_reel_confirm"],
+                    [SCISetting switchCellWithTitle:SCILocalized(@"p_reels_autonext_t") subtitle:SCILocalized(@"p_reels_autonext_s") defaultsKey:@"reels_auto_next"]
                 ]
             },
             @{